+++ src/core.js
@@ -51,3 +51,5 @@ $.extend = $.fn.extend = function(){
 
     return target;
 };
+
+if (typeof window["hideM4QVersion"] === "undefined") console.log("m4q "+$.version);