+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 1.0.3
 + [x] Events: add `$(document).ready(...)`
 + [x] Events: fix context for event handler function
++ [x] Init: add support for array. Now you can compose `$` from array of `$`
 
 ### 1.0.2
 + [x] Events: fix `function.name` property for IE11